+++ src/main/java/com/cydeo/entity/InvoiceProduct.java
@@ -19,7 +19,7 @@ public class InvoiceProduct extends BaseEntity {
 
     private int quantity;
     private BigDecimal price;
-    private int tax;
+    private BigDecimal tax;
     private BigDecimal profitLoss;
     private int remainingQuantity;
